@@ -78,7 +78,7 @@ class _ScannerScreenState extends ConsumerState<ScannerScreen>
           controller: _tabController,
           indicatorColor: Colors.white,
           labelColor: Colors.white,
-          unselectedLabelColor: Colors.white.withOpacity(0.7),
+          unselectedLabelColor: Colors.white.withValues(alpha: 0.7),
           tabs: const [
             Tab(icon: Icon(Icons.camera_alt), text: 'Camera'),
             Tab(icon: Icon(Icons.edit), text: 'Details'),
@@ -148,7 +148,7 @@ class _ScannerScreenState extends ConsumerState<ScannerScreen>
             color: Theme.of(context).colorScheme.surfaceContainerHighest,
             boxShadow: [
               BoxShadow(
-                color: Colors.black.withOpacity(0.1),
+                color: Colors.black.withValues(alpha: 0.1),
                 blurRadius: 4,
                 offset: const Offset(0, 2),
               ),
