@@ -1,18 +1,19 @@
+import 'dart:async';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:googleapis/drive/v3.dart' as drive;
 import '../core/base/base_service.dart';
 import '../core/exceptions/app_exceptions.dart';
 
 class AuthService extends BaseService {
-  final GoogleSignIn _googleSignIn = GoogleSignIn(
-    scopes: [
-      drive.DriveApi.driveFileScope,
-      'https://www.googleapis.com/auth/drive.appdata', // Required for appDataFolder access
-    ],
-  );
+  final GoogleSignIn _googleSignIn = GoogleSignIn.instance;
+  final List<String> _scopes = [
+    drive.DriveApi.driveFileScope,
+    'https://www.googleapis.com/auth/drive.appdata', // Required for appDataFolder access
+  ];
 
   GoogleSignInAccount? _currentUser;
   bool _isInitialized = false;
+  StreamSubscription<GoogleSignInAuthenticationEvent>? _authEventsSubscription;
 
   @override
   String get serviceName => 'AuthService';
@@ -28,11 +29,31 @@ class AuthService extends BaseService {
     }
 
     try {
-      // Check if user is already signed in (silent sign-in)
-      _currentUser = _googleSignIn.currentUser;
+      // Initialize the GoogleSignIn instance
+      await _googleSignIn.initialize();
 
-      // If not signed in, try to restore previous session
-      _currentUser ??= await _googleSignIn.signInSilently();
+      // Listen to authentication events to track current user
+      _authEventsSubscription = _googleSignIn.authenticationEvents.listen(
+        (GoogleSignInAuthenticationEvent event) {
+          if (event is GoogleSignInAuthenticationEventSignIn) {
+            _currentUser = event.user;
+            logInfo('User signed in via event: ${event.user.email}');
+          } else if (event is GoogleSignInAuthenticationEventSignOut) {
+            _currentUser = null;
+            logInfo('User signed out via event');
+          }
+        },
+        onError: (error) {
+          logError('Authentication event error', error);
+        },
+      );
+
+      // Try lightweight authentication (previously signInSilently)
+      final result = _googleSignIn.attemptLightweightAuthentication();
+      if (result is Future<GoogleSignInAccount?>) {
+        _currentUser = await result;
+      }
+      // If result is not a Future, we'll get updates via the event stream
 
       _isInitialized = true;
       logInfo('Auth service initialized. Signed in: $isSignedIn');
@@ -47,16 +68,28 @@ class AuthService extends BaseService {
   Future<GoogleSignInAccount?> signIn() async {
     try {
       logInfo('Initiating Google Sign-In');
-      final account = await _googleSignIn.signIn();
 
-      if (account != null) {
-        _currentUser = account;
-        logInfo('User signed in: ${account.email}');
+      // Check if platform supports authenticate method
+      if (_googleSignIn.supportsAuthenticate()) {
+        final user = await _googleSignIn.authenticate(scopeHint: _scopes);
+        _currentUser = user;
+        logInfo('User signed in: ${user.email}');
+        return user;
       } else {
-        logWarning('Google Sign-In cancelled by user');
+        // Platform doesn't support authenticate (e.g., web)
+        // For web, you would need to use renderButton from google_sign_in_web
+        throw AuthException('Platform does not support explicit authentication');
       }
-
-      return account;
+    } on GoogleSignInException catch (e) {
+      if (e.code == GoogleSignInExceptionCode.canceled) {
+        logWarning('User canceled sign-in');
+        return null;
+      }
+      logError('Google Sign-In exception', e);
+      throw AuthException(
+        'Failed to sign in: ${e.description ?? e.code.toString()}',
+        originalError: e,
+      );
     } catch (e) {
       logError('Failed to sign in', e);
       throw AuthException(
@@ -80,6 +113,11 @@ class AuthService extends BaseService {
         originalError: e,
       );
     }
+  }
+
+  /// Clean up resources
+  void cleanup() {
+    _authEventsSubscription?.cancel();
   }
 
   /// Get user email
